Fix multiple tabs opened when using Close Tab to The Right
The issue was reported at http://code.google.com/p/yjl/issues/detail?id=5

The submitted did stop tabs being opened, however Chrome also closes the
one is supposed to be the kept tab, which opened by the extension.

A timer is used to make sure that there will be a new call of keeptwo()
after Chrome closes the tab.